+++ CHANGELOG.md
@@ -7,6 +7,7 @@ Compatibility:
   - Decidim 0.29.x
 
 Features:
+  - Added application context for config constraints (whether the user is logged or not)
   - Allow to granular permissions to forced verifications.
     **NOTE** Config var `force_authorization_after_login` has changed to `force_authorizations`